Prevent scrolling when popups are open (S+ or default Schoology)
Previously old-style default Schoology popups would stay in a fixed position on the page. That behavior has been modified to prevent the user from scrolling the page. In addition, when the page is resized horizontally, the popup window is now re-centered (previously it did not move).

In addition, previously Schoology Plus modals would allow the background page to be scrolled while they were opened. This is no longer possible.

Closes #141